refactor: Service와 Retrofit 인스턴스 생성 코드 분리
